requirements.txt
streamlit
diffusers
transformers
opencv-python
# decord
open_clip_torch
torch
torchvision
Pillow
moviepy
huggingface_hub
imageio
imageio[ffmpeg]
python-dotenv
lpips
scikit-image
numpy